backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
supabase==2.5.0
httpx==0.27.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.3
pydantic-settings==2.3.1
# AI
google-generativeai==0.7.2
anthropic==0.29.0
# Job scraping
python-jobspy==1.1.75
playwright==1.44.0
# Scheduling
apscheduler==3.10.4
# Email
resend==2.2.0
# Telegram
python-telegram-bot==21.3
# PDF/DOCX parsing
PyPDF2==3.0.1
python-docx==1.1.2
# Misc
beautifulsoup4==4.12.3
tenacity==8.3.0
structlog==24.2.0
